requirements.txt
opencv-python
facenet-pytorch
torch
torchvision
numpy
pillow
flask
requests
